feat(layout): enhance UINode bounds, GodotLayout scroll constraints, and visuals

+++ src/main/kotlin/org/mdt/ui/widgets/BoxVisuals.kt
@@ -25,28 +25,28 @@ class BoxVisuals {
         NONE(0), GRAYSCALE(1), SEPIA(2), BRIGHTNESS(3), INVERT(4)
     }
 
-    /** Visual opacity (0.0 = transparent, 1.0 = fully opaque). */
-    var opacity: Double = 1.0
+    /** Visual opacity (0.0f = transparent, 1.0f = fully opaque). */
+    var opacity: Float = 1.0f
 
     /** Top-left corner radius in pixels. */
-    var topLeftRadius: Double = 8.0
+    var topLeftRadius: Float = 8f
     /** Top-right corner radius in pixels. */
-    var topRightRadius: Double = 8.0
+    var topRightRadius: Float = 8f
     /** Bottom-right corner radius in pixels. */
-    var bottomRightRadius: Double = 8.0
+    var bottomRightRadius: Float = 8f
     /** Bottom-left corner radius in pixels. */
-    var bottomLeftRadius: Double = 8.0
+    var bottomLeftRadius: Float = 8f
 
     /** Border stroke width in pixels. */
-    var borderWidth: Double = 0.0
+    var borderWidth: Float = 0f
     /** Border stroke color. */
     var borderColor: Color = Color(Color.white)
     /** Border stroke pattern style. */
     var borderStyle: BorderStyle = BorderStyle.SOLID
     /** Dash length for dashed border style. */
-    var dashLength: Double = 10.0
+    var dashLength: Float = 10f
     /** Dash to gap ratio for dashed border style. */
-    var dashRatio: Double = 0.5
+    var dashRatio: Float = 0.5f
 
     // ==========================================
     // BACKDROP BLUR PARAMETERS
@@ -58,12 +58,12 @@ class BoxVisuals {
     var blurRadius: Float = 4f
     /** Number of ping-pong blur convolution passes (1 = subtle, 4 = heavy frosted glass). */
     var blurIterations: Int = 2
-    /** Alpha blending weight of the blurred backdrop over background color (0.0 to 1.0). */
-    var backdropWeight: Double = 0.8
-    /** Blending curve exponent (0.0 to 1.0). */
-    var backdropBlend: Double = 0.8
+    /** Alpha blending weight of the blurred backdrop over background color (0.0f to 1.0f). */
+    var backdropWeight: Float = 0.8f
+    /** Blending curve exponent (0.0f to 1.0f). */
+    var backdropBlend: Float = 0.8f
     /** Minimum alpha threshold for backdrop rendering. */
-    var backdropMinAlpha: Double = 0.8
+    var backdropMinAlpha: Float = 0.8f
     /** Tint overlay color applied to the blurred backdrop texture. */
     var backdropTint: Color = Color(Color.white)
 
@@ -72,21 +72,21 @@ class BoxVisuals {
     // ==========================================
 
     var innerShadowColor: Color = Color(Pal.shadow)
-    var innerShadowSpread: Double = 0.0
-    var innerShadowBlur: Double = 0.0
+    var innerShadowSpread: Float = 0f
+    var innerShadowBlur: Float = 0f
 
     var glowColor: Color = Color(Pal.shadow)
-    var glowSpread: Double = 0.0
-    var glowBlur: Double = 0.0
+    var glowSpread: Float = 0f
+    var glowBlur: Float = 0f
 
     var filterMode: FilterMode = FilterMode.NONE
-    var filterAmount: Double = 0.0
-    var noiseAmount: Double = 0.0
+    var filterAmount: Float = 0f
+    var noiseAmount: Float = 0f
 
-    var uvScaleX: Double = 1.0
-    var uvScaleY: Double = 1.0
-    var uvOffsetX: Double = 0.0
-    var uvOffsetY: Double = 0.0
+    var uvScaleX: Float = 1.0f
+    var uvScaleY: Float = 1.0f
+    var uvOffsetX: Float = 0f
+    var uvOffsetY: Float = 0f
 
     var fillColor: Color = Color(Color.darkGray)
     var fillTexture: Texture? = null
@@ -96,12 +96,12 @@ class BoxVisuals {
     // FLUENT BUILDER HELPERS
     // ==========================================
 
-    fun radius(all: Double) {
+    fun radius(all: Float) {
         topLeftRadius = all; topRightRadius = all
         bottomLeftRadius = all; bottomRightRadius = all
     }
 
-    fun radius(tl: Double, tr: Double, br: Double, bl: Double) {
+    fun radius(tl: Float, tr: Float, br: Float, bl: Float) {
         topLeftRadius = tl; topRightRadius = tr
         bottomRightRadius = br; bottomLeftRadius = bl
     }
@@ -111,19 +111,19 @@ class BoxVisuals {
         backgroundMode = BackgroundMode.SOLID
     }
 
-    fun border(width: Double, color: Color = Color.white, style: BorderStyle = BorderStyle.SOLID) {
+    fun border(width: Float, color: Color = Color.white, style: BorderStyle = BorderStyle.SOLID) {
         borderWidth = width
         borderColor.set(color)
         borderStyle = style
     }
 
-    fun shadow(color: Color = Pal.shadow, spread: Double = 4.0, blur: Double = 8.0) {
+    fun shadow(color: Color = Pal.shadow, spread: Float = 4f, blur: Float = 8f) {
         innerShadowColor.set(color)
         innerShadowSpread = spread
         innerShadowBlur = blur
     }
 
-    fun glow(color: Color, spread: Double = 6.0, blur: Double = 12.0) {
+    fun glow(color: Color, spread: Float = 6f, blur: Float = 12f) {
         glowColor.set(color)
         glowSpread = spread
         glowBlur = blur
@@ -132,8 +132,8 @@ class BoxVisuals {
     fun backdrop(
         blur: Boolean = true,
         radius: Float = 4f,
-        weight: Double = 0.8,
-        blend: Double = 0.8,
+        weight: Float = 0.8f,
+        blend: Float = 0.8f,
         tint: Color = Color.white,
         iterations: Int = 2
     ) {
@@ -146,7 +146,7 @@ class BoxVisuals {
         this.backgroundMode = BackgroundMode.BACKDROP
     }
 
-    fun texture(tex: Texture, scaleX: Double = 1.0, scaleY: Double = 1.0) {
+    fun texture(tex: Texture, scaleX: Float = 1.0f, scaleY: Float = 1.0f) {
         fillTexture = tex
         uvScaleX = scaleX
         uvScaleY = scaleY
