+++ src/main/kotlin/org/mdt/ui/widgets/TextVisuals.kt
@@ -3,22 +3,27 @@ package org.mdt.ui.widgets
 import arc.graphics.Color
 import arc.util.Align
 
+/**
+ * ## TextVisuals
+ *
+ * Typography visual styling configuration for [TextNode].
+ */
 class TextVisuals {
     var color: Color = Color(Color.white)
     var labelAlign: Int = Align.left
     var lineAlign: Int = Align.left
     var wrap: Boolean = false
     var ellipsis: String? = null
-    var fontScaleX: Double = 1.0
-    var fontScaleY: Double = 1.0
+    var fontScaleX: Float = 1.0f
+    var fontScaleY: Float = 1.0f
     var baselineOffset: Float = 0f
 
-    fun scale(all: Double) {
+    fun scale(all: Float) {
         fontScaleX = all
         fontScaleY = all
     }
 
-    fun scale(sx: Double, sy: Double) {
+    fun scale(sx: Float, sy: Float) {
         fontScaleX = sx
         fontScaleY = sy
     }
